server/apps/school/serializers.py
from rest_framework import serializers

from server.apps.school.models import Course


class CourseSerializer(serializers.ModelSerializer):
    class Meta:
        model = Course
        fields = ['id', 'title', 'code', 'description']